Website optimization
1. index.html:
(1) Force https when it loads the font.
(2) Add media query that it loads print.css only when the medium is for
print.
(3) Move unnecessary script to the end of the file.
(4) Compress some images and force the size of images.

2. pizzeria.jpg:
Compress the size of the image.

3. main.js:
(1) Stopping FSL.
(2) Use the requestAnimationFrame to sync the file.